requirements/jenkins.txt
-r common.txt
coverage==4.0.1
django-jenkins==0.18.1
pylint==1.4.4
